test(widget): migrate inline tests to integration tests (partial)
- Migrate 55 tests from 7 widget files to tests/widget_tests.rs
- Migrated widgets: button, text, badge, checkbox, progress, spinner, divider
- Keep tests accessing private fields inline in source files
- Convert imports from use super::* to use revue::widget::*

Note: Widget module has 128 files total, migration continues
